Login and signup fully functional.
As a sub feature of the login window, the signup window is complete. This
includes adding new people to the employee table, enduring duplicate ids
arent used and error handling for missing required information or in the
event duplicate id are attempted to be used. Also created a quick use
sql file for creating the employee table. This will be converted to a
pytjon script later for auto-generation of the entire database.