Material window has been linked to main menu with some functionality.
All GUI components have been added and place within the window. There
are currently filler frames for the VOC content as well as the color
swatch. Close button has been reprogrammed to update a variable that
controls the ability of the software to open additional window. These
variables tell the main window hosting the menu when a sub-window can
or can't be opened.